Encapsulate storage class to make it more generic
Makes the storage class generic so it is possible to implement
other types of function storage and also can make it easier to abstract
it inside the tests
