+++ lib/routes.js
@@ -8,7 +8,7 @@ const Validator = require('jsonschema').Validator;
 const schemas = require('./schemas');
 const log = require('./log');
 const config = require('./config');
-const Storage = require('./storage');
+const Storage = require('./storage/storage-redis');
 const Sandbox = require('./sandbox').Sandbox;
